requirements.txt
# PyWebView 最小化演示项目依赖

# 核心依赖
pywebview>=4.4.0

# 开发工具
ruff>=0.12.0

# 可选依赖 (根据平台选择)
# Windows 平台推荐
# pythonnet>=3.0.0

# macOS 平台推荐
# pyobjc>=9.0

# Linux 平台推荐
# PyQt5>=5.15.0
# 或者
# PyQt6>=6.0.0
# 或者
# PySide2>=5.15.0
# 或者
# PySide6>=6.0.0

# 开发依赖
flake8>=6.0.0
black>=23.0.0
pyinstaller>=5.0.0